implementacion calculo de co2 general

+++ src/main/java/edu/dosw/rideci/application/port/in/Statistics/UpdateCommunityStatsUseCase.java
@@ -3,5 +3,5 @@ package edu.dosw.rideci.application.port.in.Statistics;
 import edu.dosw.rideci.application.events.TravelCompletedEvent;
 
 public interface UpdateCommunityStatsUseCase {
-    void updateCommunityStats(int co2Saved, TravelCompletedEvent event);
+    void updateCommunityStats(double co2Saved, TravelCompletedEvent event);
 }
